Backport of fix from 2.43
This is is a work-around to a crash that happens when building gcc.
Note that this version of binutils will not perform dynamic linking
correctly for i370, because it is missing the following patches
from i370-binutils master:
7a9eb9aa3ae1dae41a6cb4c5c2824c226533f2d4
a0042f85599c75657a51e1d6f9be6e5f020404db
8403340de0e593a5535bfe90bfadf3852e6e570b
303b74f9dadd7fd8dfdd0be3d0867e9e823844ce

It may also be missing newer fixes to dynamic linking.