Remove crud that will never-ever be.
Code for handling @suffix that was copied from powerpc.
It makes sense for powerpc because they have 16-bit immediates.
It does not make sense for i370 because there won't be these kinds
of fixups. That is, HLASM doesnt have them, and gcc doesn't emit
them, and there aren't any instructions that would naturally want
to make use of such markup.