Fix handling of 'BL *+6' and 'BH *+12'
These were converted to 'bc  4,*+6' and 'bc  2,*+12' where the first
number is a 4-bit mask. Forgot to check for this, and so the entire
string '4,*+6' or '2,*+12' got passed to `expression()` which ...
gave an unexpected result: neither O_constant, nor O_register, and
so the result was understood to be something that needed a fixup.

By contrast, `bc  8,.L8` worked fine, because `expression()` was
able to recognize the comma-dot-label anr returned O_constant for
the mask 8. So really, this is another example of HLASM * coming to
bite us in the weenie. So it goes. Mixing HLASM and ELF was never
a good idea, but old compilers do it, and so we're stuck with it.